refactor: make the registry the single source of model-routing truth
WHY THIS MATTERS
Model→provider routing knowledge lived in THREE places at once:

1. Each adapter's matchesModel() in the registry (e.g. Claude's
   `m.startsWith("claude")`).
2. A hard-coded prefix fallback inside providerForModel() that
   re-implemented exactly those same heuristics:

       if (model.startsWith("claude")) return "claude";
       if (model.startsWith("gemini")) return "gemini";
       if (model.startsWith("gpt-")) return "openai";

3. A static ALL_MODELS Set that resolveModel() consulted alongside the
   registry's allRegisteredModels().

Duplicated knowledge drifts. If someone changed an adapter's
matchesModel() rule (say, to also claim "anthropic.claude-*" Bedrock
IDs), the fallback in models.ts would silently disagree, and which
answer you got would depend on whether index.ts happened to be loaded
first. That is exactly the class of bug this refactor PR exists to
prevent — extensibility is only safe when routing has ONE authority.

WHAT IS "SINGLE SOURCE OF TRUTH"
A design rule: every fact in a system should have exactly one
authoritative home, and everything else should derive from it. Here the
fact is "which provider serves model X", and its home is the provider
registry: adapters declare matchesModel() and their model lists, and
all other code asks the registry instead of keeping private copies.

The duplication existed only so llmModels.test.ts could import
models.ts without loading index.ts (which registers the built-in
providers). That is a test-structure problem, and the finding's
guidance is to fix it in the tests — never by duplicating production
routing rules.

HOW THE FIX WORKS
- providerForModel() now delegates solely to findProviderForModel();
  no prefix fallback. Unknown models still throw the same
  "Unknown model id" error.
- resolveModel() now validates solely against allRegisteredModels();
  the static ALL_MODELS set is deleted.
- llmModels.test.ts imports from "../llm" (the package index) instead
  of "../llm/models". Loading index.ts registers the built-in
  providers, so the functions under test exercise the exact same
  registry path production uses. Every existing assertion — including
  "prefix-only inference without catalog validation", which now flows
  through the adapters' matchesModel() — passes unchanged.

All production consumers (userSettings.ts, chat/streaming.ts,
routes/user.ts, routes/tabular.ts) already import from "lib/llm", so
registration is guaranteed to have run before any routing call.
