+++ backend/src/lib/llm/models.ts
@@ -86,54 +86,25 @@ export const OPENCODE_GO_MESSAGES_MODEL_IDS: ReadonlySet<string> = new Set([
     "qwen3.6-plus",
 ]);
 
-// Derived (not hand-maintained) fallback set for resolveModel().
-// Built by spreading the *_MODELS arrays above, so adding a model to any
-// of those arrays automatically includes it here — no second edit site.
-//
-// Why keep this alongside allRegisteredModels()?  Two reasons:
-//   1. Test isolation: models.test.ts imports models.ts directly without
-//      importing index.ts, so no providers are registered and the registry
-//      is empty.  ALL_MODELS provides the fallback in that case.
-//   2. External providers registered via registerProvider() appear in
-//      allRegisteredModels() but NOT here — that's intentional.
-//      resolveModel() checks both, so external models are always accepted
-//      once their provider is registered.
-const ALL_MODELS = new Set<string>([
-    ...CLAUDE_MAIN_MODELS,
-    ...GEMINI_MAIN_MODELS,
-    ...OPENAI_MAIN_MODELS,
-    ...CLAUDE_MID_MODELS,
-    ...GEMINI_MID_MODELS,
-    ...OPENAI_MID_MODELS,
-    ...CLAUDE_LOW_MODELS,
-    ...GEMINI_LOW_MODELS,
-    ...OPENAI_LOW_MODELS,
-]);
-
 // ---------------------------------------------------------------------------
 // Provider inference
 // ---------------------------------------------------------------------------
+// Both functions below delegate to the provider registry — the single source
+// of truth for model→provider routing.  The registry is populated by
+// index.ts (built-in providers, on module load) and by registerProvider()
+// calls for third-party providers.  Callers must import "lib/llm" (index.ts),
+// never this file directly, so registration has always run first.
 
 /**
- * Maps a model ID to its provider string.
+ * Maps a model ID to its provider id.
  *
- * Registered providers are checked first so that externally registered
- * adapters (Ollama, Bedrock, Azure) override the built-in prefix matching
- * below — no edits to this file required to support a new provider.
- *
- * The prefix fallback keeps this function usable in test contexts that don't
- * import index.ts and therefore don't trigger provider registration.
+ * Routing is decided entirely by each registered adapter's matchesModel()
+ * (checked in registration order) — no prefix heuristics are duplicated
+ * here, so adding a provider never requires edits to this file.
  */
 export function providerForModel(model: string): string {
     const registered = findProviderForModel(model);
     if (registered) return registered.id;
-    if (model.startsWith("ollama")) return "ollama";
-    if (model.startsWith("openrouter/")) return "openrouter";
-    if (model.startsWith("vercel/")) return "vercel";
-    if (model.startsWith("opencode-go/")) return "opencode-go";
-    if (model.startsWith("claude")) return "claude";
-    if (model.startsWith("gemini")) return "gemini";
-    if (model.startsWith("gpt-")) return "openai";
     throw new Error(`Unknown model id: ${model}`);
 }
 
@@ -146,12 +117,14 @@ export const LEGACY_MODEL_IDS: Record<string, string> = {
 };
 
 /**
- * Returns id if it is a recognised model, otherwise returns fallback.
+ * Returns id if it is a model declared by any registered provider,
+ * otherwise returns fallback.
  *
- * Legacy ids are canonicalised first, then checked against the live registry
- * (which includes externally registered models) and, as a fallback for test
- * contexts where no providers have been registered, the static ALL_MODELS set
- * plus the built-in dynamic-id shapes.
+ * Legacy ids are canonicalised first: stored preferences outlive catalog
+ * renames.  Beyond the registry's declared model lists, the built-in
+ * dynamic-id shapes are accepted — Ollama tags and the three routers'
+ * namespaced catalog ids are discovered at runtime, so they never appear in
+ * any static list.
  */
 export function resolveModel(
     id: string | null | undefined,
@@ -161,7 +134,6 @@ export function resolveModel(
     if (
         canonical &&
         (allRegisteredModels().has(canonical) ||
-            ALL_MODELS.has(canonical) ||
             canonical.startsWith("ollama/") ||
             /^(?:openrouter|vercel)\/[^\s/]+\/[^\s]+$/.test(canonical) ||
             // OpenCode Go's catalog ids are single-segment ("glm-5"), not the
